+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 * **Breaking** Added additional fields for `CreateAttendeeResponse` and `CreateMeetingResponse`
 * Added `Versioning` class and `sdkVersion` API for retrieving current version of SDK
+* Added `onAudioSessionDropped` API in `AudioVideoObserver` for temporary disconnects
 
 ### Changed
 * Updated demo app to work with updated [amazon-chime-sdk-js serverless demo](https://github.com/aws/amazon-chime-sdk-js/tree/master/demos/serverless). Note that you